third_party/WebKit/LayoutTests/inspector/sources/debugger/file-system-project-mapping-expected.txt

Tests file system project mappings.


Running: testAutomaticMapping
Adding file system.
Workspace event: UISourceCodeAdded: filesystem:/var/www/html/foo.js.
Workspace event: UISourceCodeAdded: filesystem:/var/www/bar.js.
Adding network resource.
Workspace event: UISourceCodeAdded: http://localhost/html/foo.js.
Workspace event: UISourceCodeAdded: http://localhost/bar.js.
UISourceCode uri to url mappings:
    filesystem:/var/www/html/foo.js -> 
    filesystem:/var/www/bar.js -> 
Adding mapping between network and file system resources.
Emulate reloading inspector.
Workspace event: UISourceCodeAdded: filesystem:/var/www/html/foo.js.
Workspace event: UISourceCodeAdded: filesystem:/var/www/bar.js.
UISourceCode uri to url mappings:
    filesystem:/var/www/html/foo.js -> http://localhost/html/foo.js
    filesystem:/var/www/bar.js -> http://localhost/bar.js
Removing mapping between network and file system resources.
Emulate reloading inspector.
Workspace event: UISourceCodeAdded: filesystem:/var/www/html/foo.js.
Workspace event: UISourceCodeAdded: filesystem:/var/www/bar.js.
Workspace event: UISourceCodeAdded: http://localhost/html/foo.js.
Workspace event: UISourceCodeAdded: http://localhost/bar.js.
UISourceCode uri to url mappings:
    filesystem:/var/www/html/foo.js -> 
    filesystem:/var/www/bar.js -> 

Running: testScriptFileOnReloadWithDirtyFile
Adding file system.
Adding file system mapping.
Workspace event: UISourceCodeAdded: filesystem:/var/www/html/foo.js.
Workspace event: UISourceCodeAdded: filesystem:/var/www/bar.js.
Workspace event: UISourceCodeAdded: debugger:/VM1 foo.js.
Dumping uiSourceCode and script content:
    uiSourceCode: <foo content>
    script: <foo content>
  //# sourceURL=http://localhost/html/foo.js  
 
 - hasDivergedFromVM: false
 - sourceFrame._muted: false
Editing uiSourceCode:
 - hasDivergedFromVM: true
 - sourceFrame._muted: true
Committing uiSourceCode with live edit failure:
Dumping uiSourceCode and script content:
    uiSourceCode: <foo content edited>
    script: <foo content>
  //# sourceURL=http://localhost/html/foo.js  
 
 - hasDivergedFromVM: true
 - sourceFrame._muted: true
Reloading with saved but diverged uiSourceCode:
UISourceCode should not have script file after unloading scripts: true
Saved but diverged source frame should not be muted: true
Loading script.
Workspace event: UISourceCodeAdded: debugger:/VM2 foo.js.
Dumping uiSourceCode and script content:
    uiSourceCode: <foo content edited>
    script: <foo content>
  //# sourceURL=http://localhost/html/foo.js  
 
 - hasDivergedFromVM: true
 - sourceFrame._muted: true
Editing uiSourceCode again:
 - hasDivergedFromVM: true
 - sourceFrame._muted: true
Committing uiSourceCode again (with live edit success now):
Dumping uiSourceCode and script content:
    uiSourceCode: <foo content edited again>
    script: <foo content edited again>
 - hasDivergedFromVM: false
 - sourceFrame._muted: false
Reloading page:
UISourceCode should not have script file after unloading scripts: true
Saved and merged source frame should not be muted: true
Loading script.
Workspace event: UISourceCodeAdded: debugger:/VM3 foo.js.
 - hasDivergedFromVM: false
 - sourceFrame._muted: false
Editing uiSourceCode again and reloading while it is dirty:
UISourceCode should not have script file after unloading scripts: true
Dirty source frame should be muted: true
Loading script.
Workspace event: UISourceCodeAdded: debugger:/VM4 foo.js.
 - hasDivergedFromVM: true
 - sourceFrame._muted: true

Running: testRemoveProject
Adding file system.
Workspace event: UISourceCodeAdded: filesystem:/var/www/foo.js.
Dumping uiSourceCodes origin URLs:
  - file:///var/www/foo.js
Removing project:
Received project removed event.

Running: testURLAfterRenaming
Adding file system.
Adding file system mapping.
Workspace event: UISourceCodeAdded: filesystem:/var/www/html/foo.js.
Workspace event: UISourceCodeAdded: filesystem:/var/www/bar.js.
Workspace event: UISourceCodeAdded: filesystem:/var/www/foo/bar/baz.js.
Renaming files:
    uiSourceCode URL change after renaming: http://localhost/html/foo.js -> http://localhost/html/newName.js
    uiSourceCode URL change after renaming: http://localhost/bar.js -> http://localhost/newName.js
    uiSourceCode URL change after renaming: http://localhost/foo/bar/baz.js -> http://localhost/foo/bar/newName.js

Running: testProjectBasedMapping
Adding file system.
Workspace event: UISourceCodeAdded: filesystem:/var/www/html/foo.js.
Workspace event: UISourceCodeAdded: filesystem:/var/www/html2/bar.js.
(suspend state changed: true)
(suspend state changed: false)
UISourceCode uri to url mappings:
    filesystem:/var/www/html/foo.js -> http://localhost/h1/foo.js
    filesystem:/var/www/html2/bar.js -> http://localhost/h2/bar.js
(suspend state changed: true)
(suspend state changed: false)
